requirements.txt
Django==1.5.1
argparse==1.2.1
distribute==0.6.34
dj-database-url==0.2.1
django-debug-toolbar==0.9.4
django-extensions==1.1.1
django-model-utils==1.4.0
django-suit==0.2.1
ipython==0.13.2
psycopg2==2.5
six==1.3.0
wsgiref==0.1.2
